Refactor event service: Move shared DTOs and mappers to core, update repositories and queries
