+++ backend/blocking/services/event-service/src/main/java/com/ticket4u/core/Category.java
@@ -1,22 +1,22 @@
 package com.ticket4u.core;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 import java.util.List;
 
 @Getter
 @Setter
 @Entity
-@Table(name = "categories", schema = "public")
-@AllArgsConstructor
+@Builder
 @NoArgsConstructor
+@AllArgsConstructor
+@Table(name = "categories", schema = "public")
 public class Category {
+
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "categories_id_seq")
+    @SequenceGenerator(name = "categories_id_seq", sequenceName = "public.categories_id_seq", allocationSize = 1)
     private Integer id;
 
     @Column(nullable = false, unique = true, length = 255)
