+++ backend/blocking/services/event-service/src/main/java/com/ticket4u/core/projection/CategoryProjection.java
@@ -1,0 +1,10 @@
+package com.ticket4u.core.projection;
+
+/**
+ * DB projection for category-only queries returning id and name.
+ * Used for filter chips and dropdowns where event data is not needed.
+ */
+public interface CategoryProjection {
+    Integer getId();
+    String getName();
+}
