+++ backend/blocking/services/event-service/src/main/java/com/ticket4u/core/projection/CategoryWithEventProjection.java
@@ -1,13 +1,13 @@
-package com.ticket4u.feature.homepage.projection;
+package com.ticket4u.core.projection;
 
 import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
 /**
- * Projection for the single categories-with-events query.
- * Each row is a flat (category + event) pair.
- * Uses Instant because PostgreSQL JDBC driver returns Instant for timestamp columns.
+ * DB projection for flat (category + event) rows returned by the "browse by category" query.
+ * Each row represents one event; the service layer groups rows by categoryId into a map.
+ * Uses Instant because the PostgreSQL JDBC driver maps TIMESTAMPTZ columns to Instant.
  */
 public interface CategoryWithEventProjection {
     Integer getCategoryId();
