+++ backend/blocking/services/event-service/src/main/java/com/ticket4u/feature/eventdetail/service/EventDetailService.java
@@ -1,7 +1,7 @@
 package com.ticket4u.feature.eventdetail.service;
 
+import com.ticket4u.core.dto.EventSummaryResponse;
 import com.ticket4u.feature.eventdetail.dto.EventDetailResponse;
-import com.ticket4u.feature.homepage.dto.EventSummaryResponse;
 
 import java.util.List;
 import java.util.UUID;
