--- backend/blocking/services/event-service/src/main/java/com/ticket4u/feature/homepage/dto/EventSummaryResponse.java
@@ -1,16 +1,0 @@
-package com.ticket4u.feature.homepage.dto;
-
-import java.math.BigDecimal;
-import java.time.OffsetDateTime;
-import java.util.UUID;
-
-public record EventSummaryResponse(
-        UUID id,
-        String name,
-        String bannerUrl,
-        String addressLine,
-        OffsetDateTime startDate,
-        OffsetDateTime endDate,
-        BigDecimal minPrice,
-        String categoryName
-) {}
