--- backend/blocking/services/event-service/src/main/java/com/ticket4u/feature/homepage/projection/CategoryProjection.java
@@ -1,9 +1,0 @@
-package com.ticket4u.feature.homepage.projection;
-
-/**
- * Projection for category-only queries (id + name)
- */
-public interface CategoryProjection {
-    Integer getId();
-    String getName();
-}
