--- backend/blocking/services/event-service/src/main/java/com/ticket4u/feature/homepage/projection/EventWithCategoryProjection.java
@@ -1,8 +1,0 @@
-package com.ticket4u.feature.homepage.projection;
-
-/**
- * Projection for filter queries returning event data + category_name
- */
-public interface EventWithCategoryProjection extends EventSummaryProjection {
-    String getCategoryName();
-}
